refactor: migrate sync service
