+++ jsconfig.json
@@ -2,6 +2,8 @@
   "compilerOptions": {
     "baseUrl": ".",
     "jsx": "preserve",
+    "module": "esnext",
+    "moduleResolution": "bundler",
     "paths": {
       "@/*": ["./src/*"]
     }
