+++ package.json
@@ -106,6 +106,9 @@
   "homepage": "https://github.com/violentmonkey/violentmonkey",
   "license": "MIT",
   "dependencies": {
+    "@usync/drive": "^0.1.3",
+    "@usync/oauth2": "^0.1.0",
+    "@usync/sync": "^0.1.0",
     "@violentmonkey/shortcut": "^1.4.4",
     "codemirror": "^5.65.20",
     "fflate": "^0.8.3",
