+++ scripts/webpack.conf.js
@@ -42,6 +42,7 @@ const defsObj = {
     'SYNC_GOOGLE_DESKTOP_ID',
     'SYNC_GOOGLE_DESKTOP_SECRET',
     'SYNC_ONEDRIVE_CLIENT_ID',
+    'SYNC_ONEDRIVE_ACCOUNT_TYPE',
     'SYNC_DROPBOX_CLIENT_ID',
   ]),
   'process.env.INIT_FUNC_NAME': JSON.stringify(INIT_FUNC_NAME),
